refactor: coustomErros.js 파일명 customErrors.js로 오타 수정 및 그에대한 라우터 변경
